Feature/add examples (#6)
* add base examples

* fix slight bug in csv

* add all examples

* target jvm 11

* update readme

* remove examples from test reports

* fix header csv reader

* update description

* update readme

* update readme

* update the docs and the examples

* update

* test

* finished it

* remove useless line

+++ core/src/test/kotlin/io/github/pelletier197/csv/reader/TestFileUtils.kt
@@ -10,5 +10,7 @@ fun writeTestFile(content: String) {
 }
 
 fun deleteTestFile() {
-    Files.delete(Paths.get(filePath))
+    if (Files.exists(Paths.get(filePath))) {
+        Files.delete(Paths.get(filePath))
+    }
 }
